+++ CHANGELOG.md
@@ -6,15 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+* API: Portfolio - support anchor time for performance computations.
 
 ## [0.0.5] - 2020-12-31
 ### Added
 * CHANGELOG.md file.
 * Examples: Interactive example for using capon's metadata api -
   **"Analyzing the Sector-level Crash and Rebound"**.
-* API: OTC backend - screener
-* API: Nasdaq backend - extended API and screener
+* API: OTC backend - screener.
+* API: Nasdaq backend - extended API and screener.
 
 ## [0.0.4] - 2020-08-30
 ### Added