+++ CHANGELOG.md
@@ -6,10 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.6] - 2022-06-02
+### Added
 * API: Portfolio - support anchor time for performance computations.
 * API: Utils - `normalize_values`.
 * Visualization - plot history.
-* FIX: Update to new Yahoo API (user-agent).
+* Maintenance: Update to python 3.9
+### Fixed
+* Update to new Yahoo API (user-agent).
 
 ## [0.0.5] - 2020-12-31
 ### Added